requirements.txt
google-adk>=0.2.0
google-genai>=0.3.0
vertexai>=1.0.0
python-dotenv>=1.0.0
requests>=2.31.0
pandas>=2.0.0
pydantic>=2.0.0
flask>=3.0.0
flask-cors>=4.0.0
gunicorn>=21.0.0
